Added StudentManager Class
